+++ src/client.ts
@@ -1,5 +1,9 @@
 /// <reference path="types/libs.ts" />
 
+import { EventEmitter } from 'events';
+import StreamingClient from 'genesys-cloud-streaming-client';
+import StrictEventEmitter from 'strict-event-emitter-types';
+
 import {
   ISdkConfig,
   ILogger,
@@ -13,8 +17,6 @@ import {
   IUpdateOutgoingMedia,
   SdkEvents
 } from './types/interfaces';
-import StreamingClient from 'genesys-cloud-streaming-client';
-
 import {
   setupStreamingClient,
   proxyStreamingClientEvents
@@ -23,9 +25,7 @@ import { requestApi, throwSdkError, SdkError } from './utils';
 import { setupLogging } from './logging';
 import { SdkErrorTypes, SessionTypes } from './types/enums';
 import { SessionManager } from './sessions/session-manager';
-import { startMedia, startDisplayMedia } from './media-utils';
-import { EventEmitter } from 'events';
-import StrictEventEmitter from 'strict-event-emitter-types';
+import { SdkMedia } from './media/media';
 
 const ENVIRONMENTS = [
   'mypurecloud.com',
@@ -84,7 +84,8 @@ function validateOptions (options: ISdkConfig): string | null {
  */
 export class GenesysCloudWebrtcSdk extends (EventEmitter as { new(): StrictEventEmitter<EventEmitter, SdkEvents> }) {
 
-  public logger: ILogger;
+
+  logger: ILogger;
 
   readonly VERSION = '[AIV]{version}[/AIV]';
 
@@ -98,6 +99,7 @@ export class GenesysCloudWebrtcSdk extends (EventEmitter as { new(): StrictEvent
   _refreshIceServersInterval: NodeJS.Timeout;
   _config: ISdkConfig;
   sessionManager: SessionManager;
+  media: SdkMedia;
 
   get isInitialized (): boolean {
     return !!this._streamingConnection;
@@ -157,6 +159,7 @@ export class GenesysCloudWebrtcSdk extends (EventEmitter as { new(): StrictEvent
     setupLogging.call(this, options.logger);
 
     this._config.logger = this.logger;
+    this.media = new SdkMedia(this);
 
     // Telemetry for specific events
     // onPendingSession, onSession, onMediaStarted, onSessionTerminated logged in event handlers
@@ -176,7 +179,8 @@ export class GenesysCloudWebrtcSdk extends (EventEmitter as { new(): StrictEvent
    *  - guest's need a securityCode
    * @param opts optional initialize options
    */
-  public async initialize (opts?: { securityCode: string } | ICustomerData): Promise<void> {
+
+  async initialize (opts?: { securityCode: string } | ICustomerData): Promise<void> {
     let httpRequests: Promise<any>[] = [];
     if (this.isGuest) {
       let guestPromise: Promise<void>;
@@ -242,7 +246,8 @@ export class GenesysCloudWebrtcSdk extends (EventEmitter as { new(): StrictEvent
    * Start a screen share. Currently, guest is the only supported screen share.
    *  `initialize()` must be called first.
    */
-  public async startScreenShare (): Promise<{ conversationId: string }> {
+
+  async startScreenShare (): Promise<{ conversationId: string }> {
     if (this.isGuest) {
       return this.sessionManager.startSession({ sessionType: SessionTypes.acdScreenShare });
     } else {
@@ -256,37 +261,13 @@ export class GenesysCloudWebrtcSdk extends (EventEmitter as { new(): StrictEvent
    * @param roomJid jid of the conference to join. Can be made up if starting a new conference but must adhere to the format: <lowercase string>@conference.<lowercase string>
    * @param inviteeJid jid of a user to invite to this conference.
    */
-  public async startVideoConference (roomJid: string, inviteeJid?: string): Promise<{ conversationId: string }> {
+
+  async startVideoConference (roomJid: string, inviteeJid?: string): Promise<{ conversationId: string }> {
     if (!this.isGuest) {
       return this.sessionManager.startSession({ jid: roomJid, inviteeJid, sessionType: SessionTypes.collaborateVideo });
     } else {
       throwSdkError.call(this, SdkErrorTypes.not_supported, 'video conferencing not supported for guests');
     }
-  }
-
-  /**
-   * Create media with video and/or audio
-   *  `{ video?: boolean | string, audio: boolean | string }`
-   *  `true` will use the sdk default device id (or system default if no sdk default)
-   *  `string` (for deviceId) will attempt to use that deviceId and fallback to sdk default
-   * @param opts video and/or audio default device or deviceId
-   */
-  public async createMedia (opts: IMediaRequestOptions): Promise<MediaStream> {
-    if (!opts || (
-      (opts.video === undefined || opts.video === false) &&
-      (opts.audio === undefined || opts.audio === false)
-    )) {
-      throwSdkError.call(this, SdkErrorTypes.invalid_options, 'createMedia must be called with at least one media type request');
-    }
-
-    return startMedia(this, opts);
-  }
-
-  /**
-   * Creates a media stream from the screen (this will prompt for user screen selection)
-   */
-  public getDisplayMedia (): Promise<MediaStream> {
-    return startDisplayMedia();
   }
 
   /**
@@ -297,7 +278,8 @@ export class GenesysCloudWebrtcSdk extends (EventEmitter as { new(): StrictEvent
    *    - This does _not_ update the sdk `defaultOutputDeviceId`
    * @param deviceId `deviceId` for audio output, `true` for sdk default output, or `null` for system default
    */
-  public updateOutputDevice (deviceId: string | true | null): Promise<void> {
+
+  updateOutputDevice (deviceId: string | true | null): Promise<void> {
     return this.sessionManager.updateOutputDeviceForAllSessions(deviceId);
   }
 
@@ -314,7 +296,8 @@ export class GenesysCloudWebrtcSdk extends (EventEmitter as { new(): StrictEvent
    *
    * @param updateOptions device(s) to update
    */
-  public updateOutgoingMedia (updateOptions: IUpdateOutgoingMedia): Promise<void> {
+
+  updateOutgoingMedia (updateOptions: IUpdateOutgoingMedia): Promise<void> {
     if (!updateOptions ||
       (!updateOptions.stream && !updateOptions.videoDeviceId && !updateOptions.audioDeviceId)) {
       throwSdkError.call(this, SdkErrorTypes.invalid_options, 'updateOutgoingMedia must be called with a MediaStream, a videoDeviceId, or an audioDeviceId');
@@ -339,7 +322,8 @@ export class GenesysCloudWebrtcSdk extends (EventEmitter as { new(): StrictEvent
    *
    * @param options default device(s) to update
    */
-  public async updateDefaultDevices (options: IMediaDeviceIds & { updateActiveSessions?: boolean } = {}): Promise<any> {
+
+  async updateDefaultDevices (options: IMediaDeviceIds & { updateActiveSessions?: boolean } = {}): Promise<any> {
     const updateVideo = options.videoDeviceId !== undefined;
     const updateAudio = options.audioDeviceId !== undefined;
     const updateOutput = options.outputDeviceId !== undefined;
@@ -396,7 +380,8 @@ export class GenesysCloudWebrtcSdk extends (EventEmitter as { new(): StrictEvent
    * NOTE: if no `unmuteDeviceId` is provided when unmuting, it will unmute and
    *  attempt to use the sdk `defaultVideoDeviceId` as the device
    */
-  public async setVideoMute (muteOptions: ISessionMuteRequest): Promise<void> {
+
+  async setVideoMute (muteOptions: ISessionMuteRequest): Promise<void> {
     await this.sessionManager.setVideoMute(muteOptions);
   }
 
@@ -409,7 +394,8 @@ export class GenesysCloudWebrtcSdk extends (EventEmitter as { new(): StrictEvent
    *  audio stream, it will unmute and attempt to use the sdk `defaultAudioDeviceId`
    *  at the device
    */
-  public async setAudioMute (muteOptions: ISessionMuteRequest): Promise<void> {
+
+  async setAudioMute (muteOptions: ISessionMuteRequest): Promise<void> {
     await this.sessionManager.setAudioMute(muteOptions);
   }
 
@@ -417,11 +403,13 @@ export class GenesysCloudWebrtcSdk extends (EventEmitter as { new(): StrictEvent
    * Accept a pending session based on the passed in ID.
    * @param opts object with mediaStream and/or audioElement to attach to session
    */
-  public async acceptPendingSession (sessionId: string): Promise<void> {
+
+  async acceptPendingSession (sessionId: string): Promise<void> {
     await this.sessionManager.proceedWithSession(sessionId);
   }
 
-  public async rejectPendingSession (sessionId: string): Promise<void> {
+
+  async rejectPendingSession (sessionId: string): Promise<void> {
     await this.sessionManager.rejectPendingSession(sessionId);
   }
 
@@ -429,7 +417,8 @@ export class GenesysCloudWebrtcSdk extends (EventEmitter as { new(): StrictEvent
    * Accept a pending session based on the passed in ID.
    * @param opts object with mediaStream and/or audioElement to attach to session
    */
-  public async acceptSession (opts: IAcceptSessionRequest): Promise<void> {
+
+  async acceptSession (opts: IAcceptSessionRequest): Promise<void> {
     await this.sessionManager.acceptSession(opts);
   }
 
@@ -437,22 +426,41 @@ export class GenesysCloudWebrtcSdk extends (EventEmitter as { new(): StrictEvent
    * End an active session based on the session ID _or_ conversation ID (one is required)
    * @param opts object with session ID _or_ conversation ID
    */
-  public async endSession (opts: IEndSessionRequest): Promise<void> {
+
+  async endSession (opts: IEndSessionRequest): Promise<void> {
     return this.sessionManager.endSession(opts);
   }
 
   /**
    * Disconnect the streaming connection
    */
-  public disconnect (): Promise<any> {
+
+  disconnect (): Promise<any> {
     return this._streamingConnection.disconnect();
   }
 
   /**
    * Reconnect the streaming connection
    */
-  public reconnect (): Promise<any> {
+
+  reconnect (): Promise<any> {
     return this._streamingConnection.reconnect();
+  }
+
+
+  /**
+   * Ends all active sessions, disconnects the 
+   *  streaming-client, removes all event listeners, 
+   *  and cleans up media.
+   * 
+   * WARNING: calling this effectively renders this SDK
+   *  instance useless. A new instance will need to be 
+   *  created after this is called. 
+   */
+  async destroy (): Promise<any> {
+    // TODO: lots of things
+    this.removeAllListeners();
+    this.media.destory();
   }
 
   async _refreshIceServers () {