+++ src/sessions/base-session-handler.ts
@@ -1,11 +1,19 @@
+import { JingleReason } from 'stanza/protocol';
+
 import { GenesysCloudWebrtcSdk } from '../client';
 import { LogLevels, SessionTypes, SdkErrorTypes } from '../types/enums';
 import { SessionManager } from './session-manager';
-import { IPendingSession, IStartSessionParams, IAcceptSessionRequest, ISessionMuteRequest, IExtendedMediaSession, IUpdateOutgoingMedia, IJingleReason } from '../types/interfaces';
-import { checkHasTransceiverFunctionality, hasOutputDeviceSupport, logDeviceChange, startMedia } from '../media-utils';
+import { checkHasTransceiverFunctionality, logDeviceChange } from '../media/media-utils';
 import { throwSdkError } from '../utils';
 import { ConversationUpdate } from '../types/conversation-update';
-import { JingleReason } from 'stanza/protocol';
+import {
+  IPendingSession,
+  IStartSessionParams,
+  IAcceptSessionRequest,
+  ISessionMuteRequest,
+  IExtendedMediaSession,
+  IUpdateOutgoingMedia
+} from '../types/interfaces';
 
 type ExtendedHTMLAudioElement = HTMLAudioElement & {
   setSinkId (deviceId: string): Promise<undefined>;
@@ -92,7 +100,7 @@ export default abstract class BaseSessionHandler {
       params
     };
     const outputDeviceId = this.sdk._config.defaults.outputDeviceId || '';
-    const isSupported = hasOutputDeviceSupport();
+    const isSupported = this.sdk.media.getState().hasOutputDeviceSupport;
 
     /* if we have an audio element _and_ are in a supported browser */
     if (session._outputAudioElement && isSupported) {
@@ -208,7 +216,7 @@ export default abstract class BaseSessionHandler {
 
     if (!stream) {
       try {
-        stream = await startMedia(this.sdk, {
+        stream = await this.sdk.media.startMedia({
           audio: options.audioDeviceId,
           /* if video is muted, we don't want to request it */
           video: !session.videoMuted && options.videoDeviceId,